Added refactored files to Github Repo so that setup can be reproduced.

+++ projects/audio/Cloners/getVoiceCloners.py
@@ -12,17 +12,18 @@ cloners = {
 def getVoiceCloners():
 	for cloner in cloners.keys(): 
 		print(f'Getting the {cloners[cloner]["name"]} Cloner...')
-		if(cloners[cloner]["batscript"]):
-			os.system(f'setup-{cloners[cloner]["repo"]}.bat')
-		else:
-			if(not(os.path.exists(cloners[cloner]["repo"]))):
-				os.system(f'git clone {cloners[cloner]["link"]} {cloners[cloner]["repo"]}')
-			os.chdir(f'{cloners[cloner]["repo"]}')	
-			if(os.path.exists('requirements.txt')):
-				os.system(f'pip install -r requirements.txt')
-			if(os.path.exists('setup.py')):
-				os.system(f'python setup.py install')
-			os.chdir('..')
+		if(not(os.path.exists(cloner))):
+			if(cloners[cloner]["batscript"]):
+				os.system(f'setup-{cloners[cloner]["repo"]}.bat')
+			else:
+				if(not(os.path.exists(cloners[cloner]["repo"]))):
+					os.system(f'git clone {cloners[cloner]["link"]} {cloners[cloner]["repo"]}')
+				os.chdir(f'{cloners[cloner]["repo"]}')	
+				if(os.path.exists('requirements.txt')):
+					os.system(f'pip install -r requirements.txt')
+				if(os.path.exists('setup.py')):
+					os.system(f'python setup.py install')
+				os.chdir('..')
 
 # Hopefully this only has to be run once.
 if __name__ == '__main__':
